Remove sampling usage bit requirement for output buffers
Skia RenderEngine hits a different path for creating SkSurfaces that
does not enforce presence of the sampling bit, so we don't really need
to enforce it here.

Note that in practice the sampling bit may still be provided in
anticipation of SF caching, so this is really just for tests/cleanup.

Bug: 180650640
Test: First, patch out sampling usage bit in RenderSurface
Test: boots
Test: simulate secondary displays
Test: force GPU composition
Change-Id: I35558beec60b1c228e1fa0ad6382482869e1719c