Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > > Patch Set 4:
> > > 
> > > Huh, not sure about your case, but Xiaomi's impl which is also always on HBM works just fine w/o this
> > > You could do it in kernel anyways
> > 
> > im not really sure how your always on HBM would work without this, unless you dont use fod extension and just trigger HBM with a sysfs node.
> 
> No, Xiaomi enable HBM and icon layer w one z value, then pressed layer w another

samsung uses their usageBits to only trigger HBM on or off and they use framework dimming right when the system request fingerprint biometric (ex. right when you turn on the screen). This is needed so that both untouched layer and touched layer would still be HBM turned on.

Patch-set: 4