Load ANGLEResetDisplayPlatform via dlsym
... which is slightly more defensive than loading the function via
eglGetProcAddress. When loading the function via eglGetProcAdress,
the returned function pointer will be one of the ext forwardering
helpers (__glExtFwd0 in this observed case). The process can then
crash if eglTerminate() is called on a new thread that has not yet
called eglMakeCurrent() where the thread local gl_hooks_t is 0x0.

Bug: b/181162754
Test: HeifWriterTest on Cuttlefish w/ SwANGLE
Change-Id: Ia8e55c899f9fde91215ce75313b33517f8440fee