Trace DisplayFrame present time as the slice end time
To find the present time of a frame, we'd have to go to hw vsync
track and check for the fence signal time. To make things easier, this
change traces the DisplayFrame's present time as the slice's end time.
This also makes it inline with the expected track which uses present
time (end and present are same for sf predictions).

Bug: 180423820
Test: libsurfaceflinger_unittest
Change-Id: I2359417337eb08cde0adb881101087fa37d4e6f2