Change fps listener permission to allow systemui to use it.
Require READ_FRAME_BUFFER permission instead of ACCESS_SURFACE_FLINGER,
so that systemui is allowed to listen to FPS measurements.

Strictly speaking, the fps listener does not require as strict of
permissions as screenshots because there's no on-screen content, but
systemui is already requesting this permission, and we probably don't
want arbitrary apps to backdoor into this api.

Bug: 174956756
Test: end to end integration test to check that systemui does not crash
when registering the listener.

Change-Id: I9149618950988250ae2649ac5242d5165b5037db