libgui: add aidl header lib
This allows vendors to include the AIDL headers into their builds, since
otherwise if they try to link against a libgui stub file they run into
linker errors as the AIDL headers are not included in libgui_headers.
Furthermore, because AIDL definitions are generated at compile-time,
they cannot be placed into libgui_headers directly because APEX modules
depend on libgui_headers, but because the AIDL definitions are unstable
this is inherentrly problematic.

Bug: 181034286
Test: builds
Change-Id: I3d5c14b6b9cec697d06c109375f902ebc83fe726