Implement SkiaGLRenderEngine::primeCache
Bug: 178661709
Test: perfetto/systrace

SkiaGLRenderEngine is unable to write to disk, so it cannot have a
persistent cache for shaders. Instead, warm up the cache by calling
drawLayers in the ways we expect to see it called during actual use.

Currently, this just uses parameters seen when opening the dialer, which
required compiling several shaders. Future CLs will expand this to cover
other cases.

Remove the const modifier from primeCache. Conceptually, it is not
const, since we want to change some state (what's in the cache). More
practically, it is now calling drawLayers, which is not const.

Change-Id: I4bad4ff6cf79d4cd81eef1a5bee6530b3bdc0e84