SF: check isVsyncValid based on the transaction submitter
isVsyncValid is used to throttle applications to a lower frame rate than
the display refresh rate. With the current implementation of BlastSync
we may get a single transaction that contains layers from different
applications, we may cause to throttle unindended applications
(such as SystemUI). Instead, we are checking the transaction submitter
for whether we should throttle the transaction or not.

Bug: 182587984
Test: atest FrameRateOverrideHostTest
Change-Id: I7dd0c4c1ac945d6142afcdc8ba62c5a48a6dbb00