Move drag event to InputDispatcher (7/n)
This CL handle the special case while using a stylus device. During drag
and drop and the primary button was down, we will check the button being
lifted in the future, if that happens, we'll drop the item.

This includes the behavior:
-- Touch screen with stylus -> start drag drop with button pressed ->
   move and drag item -> drop item by releasing button.

We could still drop item when receiving ACTION_UP.

Bug: 19621008
Bug: 158242495
Test: atest inputflinger_tests
Change-Id: Ia84f4fd85367f96272af3a19c3db50c4216de1b0