SkiaRE: respect disableBlending
Fixes: 182617606
Test: atest librenderengine_test
      RenderEngineTest#testClear
      RenderEngineTest#testDisableBlendingBuffer

When drawing an image or a solid color, use SkBlendMode::kSrc to
avoid blending with other layers. Assert that disableBlending is not
combined with a shadow, since our existing API does not support that.
Continue to ignore disableBlending with blur, which we do not expect
to see as input.

Add RenderEngineTest#testClear and #testDisableBlendingBuffer to verify
the fix.

Change-Id: Ie12cbf497ecb6b71a9b4b625710161fce2b06ee4