Clear transaction that's passed to mergeWithNextTransaction
The std::move wasn't actually clearing the incoming transaction so it
was possible for the transaction to get applied by something else. This
would cause unexpected behavior since the caller intented for the
transaction to be merged with the next frame.

Instead, just explicitly clear the transaction after storing in the
vector.

Test: No longer flickers when resizing pip
Fixes: 182985130
Change-Id: Ie6ed0825074fb51ae3c260832002c012606241a1