RefreshRateConfigs: Fix for default seamlessness and seamed layers
When there's a focused layer with has requested a seamless change
a layer with default seamlessness is not supposed to be able to
change the refresh rate group. This behaviour is covered by the test
groupSwitchingWithTwoLayersDefaultFocusedAndSeamed.

However if the seamed layer is not focused a default layer is
allowed to switch back to the group of the default mode.

Test: atest RefreshRateConfigsTest
Bug: 183128900
Change-Id: I1a7da19fbfda3b221bda6f2d2dbb9fa65de23cf5