libbinder_ndk: document #apex service caveats
Code in an APEX may be run in the same process as a platform code and
have permissions to get any service. While this is targetted to be fixed
in b/139325195, it should be documented for now.

When these services are used from within the VNDK, binder's vendor
stability mechanism prevents misuse of this API, except inside of the
product partition APEX. This is tracked in b/182526613. However, it is
not called out here because due to their nature, these libraries aren't
expected to register services or have dependencies on vendor.

Bug: 181551684
Test: N/A

Change-Id: I0ee551b731821fb114905e5f83fc7f81ef8cd395