Remove Flattener::mLayersHash
Bug:163076219
Test: make

This field is computed and then immediately returned by another method.
But it is never read again until it has been rewritten. The client of
Flattener::flattenLayers stores the value, but this class doesn't need
to store it, too. Remove the field, and switch updateLayersHash to
computeLayersHash, which simply computes and returns the value. This is
simpler and clearer.

Change-Id: I065f1968bbe66e94180c57c54ba4f32a3a8ddf40