libbinder: vsock support for RPC
This adds support for binder RPC to be used with virtual sockets.

One decision made here - should RpcConnection.h take a sockaddr object,
or not? I've chosen not to expose sockaddr directly because it is a
low-level type (specifically for unix domain sockets, where the path may
not be null-terminated).

Fixes: 182938272
Test: binderRpcTest
Change-Id: I29bbbd2745f71d8ed7d4b8f8273fce28de0a9779