Store a wp GraphicBuffer pointer instead of an sp on LayerState
LayerStates may be persistently stored as part of an example layer
stack in the predictor, so refrain from storing a ref-counted pointer
and instead store the weak pointer. The GraphicBuffer is only used as
part of testing the differences between two LayerStates to check if
there was a buffer update, so we don't really need a strong reference.

Also clean up some hashing and difference field computations, since the
buffer was never really used for those so there's no reason to
parameterize the skip fields.

Bug: 183977995
Test: builds, boots
Test: Observe no increase in ION meminfo after experimenting with device
Change-Id: Ibfc74b4e794596bb8bb80487154522cf53a71e31