Use zetasketch for input statistics
Switch LatencyAggregator to use the actual zetasketch library. Report
the data to westworld via 2 atoms. 1 atom is for slow events, and 1 is
for latency sketches.

Estimating < 25 Kb per atom after few minutes of non-stop interaction with the screen.
Added a hard limit at 20 Kb

Dumped the data to a file on the device, and then read that data using
the full zetasketch implementation. Confirmed that the data looks OK.

Bug: 167947340
Test: adb shell dumpsys input and review the size of the resulting
sketches

Change-Id: I69b0f3753c86226886186d5f6edd779d2f9404e5