Remove usages of the DISPLAY_JANK_REPORTED atom
This was originally experimental code for Android 11, but we haven't
used this and collection is not enabled.

This also won't work well on Pixel devices because of complicated
apex-reasons; SurfaceFlinger starts up before apex directories are
mounted which means that SurfaceFlinger cannot depend on any apex
modules, and we won't be able to resolve this dependency in Android 12.

Bug: 184393731
Test: builds

Change-Id: I563d41f38dd2d233f041dc52f823f12b4f5c5930