Delete translation service feature files.
Currently, the Translation uses feature to determine if we need to
start the translation system service. But our case is like the
content capture not the autofill, the autofill can have the feature
without any service defined and the user can install one later. But
translation isn't, we should need to check config instead of
feature.

Bug: 183360041
Test: atest CtsTranslationTestCases
Change-Id: Ic44f085a688bcd9dcb6582fd5e81531eea22ab2f