SurfaceFlinger: Peek in current state for transform hint
Imagine a sequence like this:
1. Apply transaction to rotate display
2. Create surface
3. Transaction to rotate display processed

From a client perspective we executed 2 after 1 and would expect it
to observe the new transform hint, but it may not due so
due to the waiting for current->drawing state swap. Since we hold
the lock on the binder thread we can peek in the current state and
return the most recent hint.

Bug: 184584284
Test: Existing tests pass
Change-Id: Ic4927f517757c3c3dcfc1570e66b40ed370496de