Remove hwcDuration from FrameTimeline
We added hwcDuration previously to give some fixed threshold for
composer. However, this isn't accurate or useful since the calls to
client composition are not blocking on the GPU and hence a fixed
duration threshold will not work there. Instead, we will consider the
end time for SF as the time postComposition is called. This removes any
ambiguity and just takes the whole CPU time (excluding postComposition)
that SF ran for the frame.

Bug: 169876734
Test: libsurfaceflinger_unittest
Change-Id: I0e8f0b15f82d8baa884db5e9177c5696257fe11f