SurfaceFlinger::getNewTexture: Don't deadlock on main thread
This codepath may be invoked on the main thread via the
RefreshRateOverlay surface creation codepath. In such case it will
trigger a deadlock as we wait on the main thread to complete while
blocking the main thread. This can be observed by running
RefreshRateOverlayTests with ~5 iterations. It seems safest to just
avoid this deadlock.

Bug: 184991996
Test: Existing tests pass
Change-Id: I3c5c495b8e93761813157840c9878c73bcdee579