Remove surface damage from planner LayerState
Confirmed that the devices that we're prototyping this feature on don't
need to change their composition strategy based on the surface damage,
which means that this should not be included in the LayerState.

This also fixes an issue where when idle fallback is enabled in
SurfaceFlinger, the cached sets are always decomposed because the
surface damage becomes empty for every layer, which tricked the planner
into thinking the layer stack geometry changed.

Bug: 184735459
Test: Checked that cached sets are generated for static screen
workloads.

Change-Id: Idc0cd220bd31b932bb5f2421701790d52b10dfbc