SF: fix getFrameRateForLayerTree
In commit bdda8f08e38620b071d1c77e8345b141f7fcedfe we removed the
wp<Layer> from LayerHistory and instead updated LayerHistory when
something changed in the Layer itself. That change didn't take into
account that the FrameRateForLayerTree for a layer might change
as a result of a setFrameRate called on a different Layer. In this change
we update the FrameRateForLayerTree for each layer that is impacted
by a setFrameRate call and update LayerHistory accordingly.

Test: atest SetFrameRateTest
Bug: 169873384
Bug: 183225713
Change-Id: Idf044e82225697e6132200dee9cec45b08fe4f5f