Fix crop scaling with BSL
With BQL, each layer had a concept of buffer space and layer space.
The caller could set a buffer crop and a layer crop. The buffer crop
would be applied in buffer space and the content would be scaled
to the buffer size. Then the scaled content would then be scaled
again to the layer size (also referred to as the window size).

With BSL we do not have a concept of buffer space and layer space.
So we should not set a buffer/content crop otherwise this might
result in incorrect scaling or cropping. To fix this, in BBQ, when
calculating the scale, use the buffer crop provided by the client
and in SurfaceFlinger, instead of setting a buffer crop only set
the layer crop. The buffer crop/content crop can be cleaned up
once BQL is removed.

Test: BLASTBufferQueueTest (cropped buffer scales to buffer and window size)
Test: 720p content in android tv is not cropped incorrectly
Bug: 178622186
Change-Id: I173df901120a43f397f6d623a7e7b93537a508e2